add repos detail model
